+++ server/src/main/java/io/spine/server/aggregate/Aggregate.java
@@ -36,8 +36,8 @@ import io.spine.server.aggregate.model.AggregateClass;
 import io.spine.server.aggregate.model.Applier;
 import io.spine.server.command.CommandHandlingEntity;
 import io.spine.server.command.model.CommandHandlerMethod;
-import io.spine.server.entity.BatchDispatch;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.BatchDispatch;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.EventPlayer;
 import io.spine.server.event.EventReactor;
 import io.spine.server.event.model.EventReactorMethod;
