+++ server/src/main/java/io/spine/server/aggregate/AggregateTestSupport.java
@@ -23,7 +23,7 @@ package io.spine.server.aggregate;
 import com.google.common.annotations.VisibleForTesting;
 import io.spine.annotation.Internal;
 import io.spine.security.InvocationGuard;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.server.type.EventEnvelope;
 import io.spine.server.type.MessageEnvelope;
