+++ server/src/main/java/io/spine/server/aggregate/EventImportEndpoint.java
@@ -22,8 +22,8 @@ package io.spine.server.aggregate;
 
 import io.spine.core.Event;
 import io.spine.logging.Logging;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.type.EventEnvelope;
 
 /**
