+++ server/src/main/java/io/spine/server/command/AbstractCommander.java
@@ -32,8 +32,8 @@ import io.spine.server.command.model.CommandReactionMethod;
 import io.spine.server.command.model.CommandSubstituteMethod;
 import io.spine.server.command.model.CommanderClass;
 import io.spine.server.commandbus.CommandBus;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.event.EventDispatcherDelegate;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.CommandEnvelope;
