+++ server/src/main/java/io/spine/server/command/CommandHandlingEntity.java
@@ -28,7 +28,7 @@ import io.spine.change.MessageMismatch;
 import io.spine.change.StringMismatch;
 import io.spine.change.ValueMismatch;
 import io.spine.protobuf.ValidatingBuilder;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.TransactionalEntity;
 import io.spine.server.type.CommandEnvelope;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
