+++ server/src/main/java/io/spine/server/entity/EventPlayingTransaction.java
@@ -25,6 +25,7 @@ import com.google.protobuf.Message;
 import io.spine.annotation.Internal;
 import io.spine.core.Version;
 import io.spine.protobuf.ValidatingBuilder;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.event.EventDispatch;
 import io.spine.server.type.EventEnvelope;
