+++ server/src/main/java/io/spine/server/entity/Phase.java
@@ -23,6 +23,7 @@ package io.spine.server.entity;
 import io.spine.annotation.Internal;
 import io.spine.core.Signal;
 import io.spine.core.SignalId;
+import io.spine.server.dispatch.DispatchOutcome;
 
 import static com.google.common.base.Preconditions.checkNotNull;
