+++ server/src/main/java/io/spine/server/event/AbstractEventReactor.java
@@ -32,8 +32,8 @@ import io.spine.logging.Logging;
 import io.spine.protobuf.TypeConverter;
 import io.spine.server.BoundedContext;
 import io.spine.server.ContextAware;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.ProducedEvents;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.ProducedEvents;
 import io.spine.server.event.model.EventReactorClass;
 import io.spine.server.event.model.EventReactorMethod;
 import io.spine.server.integration.ExternalMessageClass;
