+++ server/src/main/java/io/spine/server/event/AbstractEventSubscriber.java
@@ -30,8 +30,8 @@ import io.spine.logging.Logging;
 import io.spine.server.BoundedContext;
 import io.spine.server.ContextAware;
 import io.spine.server.bus.MessageDispatcher;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Ignore;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Ignore;
 import io.spine.server.event.model.EventSubscriberClass;
 import io.spine.server.event.model.SubscriberMethod;
 import io.spine.server.integration.ExternalMessageClass;
