+++ server/src/main/java/io/spine/server/event/EventDispatch.java
@@ -21,7 +21,7 @@
 package io.spine.server.event;
 
 import io.spine.annotation.Internal;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.TransactionalEntity;
 import io.spine.server.type.EventEnvelope;
