+++ server/src/main/java/io/spine/server/model/AbstractHandlerMethod.java
@@ -27,8 +27,8 @@ import io.spine.base.Error;
 import io.spine.base.ThrowableMessage;
 import io.spine.core.MessageId;
 import io.spine.core.Signal;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.model.declare.ParameterSpec;
 import io.spine.server.type.MessageEnvelope;
 import io.spine.type.MessageClass;
