+++ server/src/main/java/io/spine/server/model/CommandProducingMethod.java
@@ -26,8 +26,8 @@ import io.spine.client.ActorRequestFactory;
 import io.spine.client.CommandFactory;
 import io.spine.core.ActorContext;
 import io.spine.core.Command;
-import io.spine.server.entity.ProducedCommands;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.ProducedCommands;
+import io.spine.server.dispatch.Success;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.MessageEnvelope;
 import io.spine.type.MessageClass;
