+++ server/src/main/java/io/spine/server/model/EventProducingMethod.java
@@ -25,8 +25,8 @@ import io.spine.base.EventMessage;
 import io.spine.core.Event;
 import io.spine.core.Version;
 import io.spine.server.EventProducer;
-import io.spine.server.entity.ProducedEvents;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.ProducedEvents;
+import io.spine.server.dispatch.Success;
 import io.spine.server.event.EventFactory;
 import io.spine.server.type.EventClass;
 import io.spine.server.type.MessageEnvelope;
