+++ server/src/main/java/io/spine/server/model/HandlerMethod.java
@@ -22,8 +22,8 @@ package io.spine.server.model;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import com.google.errorprone.annotations.Immutable;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.type.MessageEnvelope;
 import io.spine.type.MessageClass;
