+++ server/src/main/java/io/spine/server/procman/PmEventEndpoint.java
@@ -22,7 +22,7 @@ package io.spine.server.procman;
 
 import io.spine.annotation.Internal;
 import io.spine.server.delivery.EventEndpoint;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.type.EventEnvelope;
 
 /**
