+++ server/src/main/java/io/spine/server/procman/PmTransaction.java
@@ -27,11 +27,11 @@ import io.spine.core.Event;
 import io.spine.core.Version;
 import io.spine.protobuf.ValidatingBuilder;
 import io.spine.server.command.DispatchCommand;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.entity.CommandDispatchingPhase;
-import io.spine.server.entity.DispatchOutcome;
 import io.spine.server.entity.EventDispatchingPhase;
 import io.spine.server.entity.Phase;
-import io.spine.server.entity.Success;
 import io.spine.server.entity.Transaction;
 import io.spine.server.entity.VersionIncrement;
 import io.spine.server.event.EventDispatch;
