+++ server/src/main/java/io/spine/server/projection/Projection.java
@@ -25,8 +25,8 @@ import io.spine.annotation.Internal;
 import io.spine.base.Error;
 import io.spine.core.Event;
 import io.spine.protobuf.ValidatingBuilder;
-import io.spine.server.entity.BatchDispatch;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.BatchDispatch;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.EventPlayer;
 import io.spine.server.entity.HasVersionColumn;
 import io.spine.server.entity.TransactionalEntity;
