+++ server/src/test/java/io/spine/server/aggregate/AggregateTest.java
@@ -45,8 +45,8 @@ import io.spine.server.aggregate.given.aggregate.TaskAggregateRepository;
 import io.spine.server.aggregate.given.aggregate.TestAggregate;
 import io.spine.server.aggregate.given.aggregate.TestAggregateRepository;
 import io.spine.server.commandbus.CommandBus;
-import io.spine.server.entity.BatchDispatch;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.BatchDispatch;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.server.type.EventClass;
