+++ server/src/test/java/io/spine/server/aggregate/AggregateTransactionTest.java
@@ -22,7 +22,7 @@ package io.spine.server.aggregate;
 import com.google.protobuf.Message;
 import io.spine.core.Event;
 import io.spine.core.Version;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.Transaction;
 import io.spine.server.entity.TransactionListener;
 import io.spine.server.entity.TransactionTest;
