+++ server/src/test/java/io/spine/server/command/model/CommandReactionMethodTest.java
@@ -29,8 +29,8 @@ import io.spine.server.command.model.given.reaction.ReOneParam;
 import io.spine.server.command.model.given.reaction.ReOptionalResult;
 import io.spine.server.command.model.given.reaction.ReTwoParams;
 import io.spine.server.command.model.given.reaction.TestCommandReactor;
-import io.spine.server.entity.DispatchOutcome;
-import io.spine.server.entity.Success;
+import io.spine.server.dispatch.DispatchOutcome;
+import io.spine.server.dispatch.Success;
 import io.spine.server.event.EventReceiver;
 import io.spine.server.type.EventEnvelope;
 import io.spine.test.command.CmdAddTask;
