+++ server/src/test/java/io/spine/server/projection/ProjectionTransactionTest.java
@@ -23,7 +23,7 @@ import com.google.protobuf.Message;
 import io.spine.core.Event;
 import io.spine.core.Version;
 import io.spine.core.Versions;
-import io.spine.server.entity.DispatchOutcome;
+import io.spine.server.dispatch.DispatchOutcome;
 import io.spine.server.entity.Transaction;
 import io.spine.server.entity.TransactionListener;
 import io.spine.server.entity.TransactionTest;
